Backend/requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers

numpy
tensorflow
Pillow
opencv-python

fastapi
uvicorn
pydantic
python-dotenv
requests